+++ configurer/core/models/wazuh_config_mapping.py
@@ -44,11 +44,13 @@ class WazuhConfigMapping:
         for file in files_config:
             if (file_content := file.get(type, None)) and file.get("path", None):
                 try:
-                    content.append({
-                        "path": Path(file["path"]),
-                        "keys": file_content["keys"],
-                        "values": file_content["values"],
-                    })
+                    content.append(
+                        {
+                            "path": Path(file["path"]),
+                            "keys": file_content["keys"],
+                            "values": file_content["values"],
+                        }
+                    )
                 except KeyError as err:
                     raise KeyError(f"Missing 'keys' or 'values' key in '{type}' mapping file section: {err}") from err
             else: